Move gw worklow out
