+++ src/nomad_simulations/schema_packages/workflow/__init__.py
@@ -1,7 +1,8 @@
-from .beyond_dft import BeyondDFTResults, DFTGWWorkflow, DFTTBWorkflow
+from .beyond_dft import BeyondDFTResults, DFTTBWorkflow
 from .equation_of_state import EquationOfState
 from .general import ParallelWorkflow, SerialWorkflow, SimulationWorkflow
 from .geometry_optimization import GeometryOptimization
+from .gw import DFTGWWorkflow
 from .molecular_dynamics import MolecularDynamics
 from .phonon import Phonon
 from .photon_polarization import PhotonPolarizationWorkflow
